clean up backend directories
