+++ backend/routes/authRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const authController = require('../controllers/authController');
+
+router.post('/signup', authController.signup);
+router.post('/login', authController.login);
+
+module.exports = router;
